Changed radius from int to double

+++ lib/src/util.dart
@@ -90,7 +90,7 @@ types.User processUserDocument(DocumentSnapshot<Map<String, dynamic>> doc) {
   final isBijlesgever = doc.data()?['isBijlesgever'] as bool?;
   final metadata = doc.data()?['metadata'] as Map<String, dynamic>?;
   final locatie = doc.data()?['locatie']['geopoint'] as GeoPoint?;
-  final radius = doc.data()?['radius'] as int?;
+  final radius = doc.data()?['radius'] as double?;
   final fcm = doc.data()?['fcm'] as Map<String, dynamic>;
   final Map<String, DateTime> fcmDoc =
       fcm.map((k, v) => MapEntry(k, (v as Timestamp).toDate()));
